+++ zproject/urls.py
@@ -177,6 +177,7 @@ from zerver.views.user_groups import (
     delete_user_group,
     edit_user_group,
     get_is_user_group_member,
+    get_subgroups_of_user_group,
     get_user_group,
     get_user_group_members,
     update_subgroups_of_user_group,
@@ -379,7 +380,11 @@ v1_api_and_json_patterns = [
         GET=get_user_group_members,
         POST=update_user_group_backend,
     ),
-    rest_path("user_groups/<int:user_group_id>/subgroups", POST=update_subgroups_of_user_group),
+    rest_path(
+        "user_groups/<int:user_group_id>/subgroups",
+        POST=update_subgroups_of_user_group,
+        GET=get_subgroups_of_user_group,
+    ),
     rest_path(
         "user_groups/<int:user_group_id>/members/<int:user_id>", GET=get_is_user_group_member
     ),